Rename Format3Builder to Builder
